+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/init/ItemModels.java
@@ -243,7 +243,7 @@ public class ItemModels extends ItemModelProvider {
 	}
 
 	void unionApprenticeArmor(String name) {
-		String slot = name.substring("apprentice_".length());
+		String slot = name.substring(name.lastIndexOf('_') + 1);
 		getBuilder(name).parent(new ModelFile.UncheckedModelFile("item/generated")).texture("layer0", "minecraft:item/leather_" + slot);
 	}