+++ src/main/java/online/kingdomkeys/kingdomkeys/datagen/init/LanguageESES.java
@@ -1375,9 +1375,18 @@ public class LanguageESES extends KKLanguageProvider {
         //Armour
         add("gui.summonarmor.notenoughspace", "No tienes suficiente espacio en el inventario");
 
-        addItem(ModItems.apprentice_Chestplate, "Túnica de aprendiz");
-        addItem(ModItems.apprentice_Leggings, "Pantalones de aprendiz");
-        addItem(ModItems.apprentice_Boots, "Botas de aprendiz");
+        addItem(ModItems.apprentice1_Chestplate, "Túnica de aprendiz I");
+        addItem(ModItems.apprentice1_Leggings, "Pantalones de aprendiz I");
+        addItem(ModItems.apprentice1_Boots, "Botas de aprendiz I");
+        addItem(ModItems.apprentice2_Chestplate, "Túnica de aprendiz II");
+        addItem(ModItems.apprentice2_Leggings, "Pantalones de aprendiz II");
+        addItem(ModItems.apprentice2_Boots, "Botas de aprendiz II");
+        addItem(ModItems.apprentice3_Chestplate, "Túnica de aprendiz III");
+        addItem(ModItems.apprentice3_Leggings, "Pantalones de aprendiz III");
+        addItem(ModItems.apprentice3_Boots, "Botas de aprendiz III");
+        addItem(ModItems.apprentice4_Chestplate, "Túnica de aprendiz IV");
+        addItem(ModItems.apprentice4_Leggings, "Pantalones de aprendiz IV");
+        addItem(ModItems.apprentice4_Boots, "Botas de aprendiz IV");
 
         addItem(ModItems.organizationRobe_Helmet, "Capucha de la Organización");
         addItem(ModItems.organizationRobe_Chestplate, "Abrigo de la Organización");